requirements-reconstruction.txt
# 3D Reconstruction and Mesh Generation dependencies
open3d<0.16.0  # Latest version compatible with Python 3.7
numpy-stl<3.0.0
# pymeshlab<2022.2.0  # Removed due to compatibility issues with Python 3.7
trimesh<3.15.0
pyntcloud<0.2.0  # For point cloud processing
# pyvista<0.35.0   # Removed due to compatibility issues with Python 3.7

# Additional visualization for reconstruction
plotly<5.14.0
seaborn<0.12.0 